+++ CHANGELOG.md
@@ -25,6 +25,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
   [GEOD-1229](https://opensource.ncsa.illinois.edu/jira/browse/GEOD-1229)
 - Zoom to the features when drawing map for the first time
   [GEOD-1223](https://opensource.ncsa.illinois.edu/jira/browse/GEOD-1223)
+- Use parameter notation for detail page URL parameters
+  [GEOD-1222](https://opensource.ncsa.illinois.edu/jira/browse/GEOD-1222)
 
 ### Fixed
 - Ensure the map pans correctly when switching Sensors on the Explore Page